LayoutTests/platform/glib/imported/w3c/web-platform-tests/editing/plaintext-only/paste.https_white-space=normal-expected.txt

abc
def
AabcB

PASS Pasting text in <b>
PASS Pasting text in <span>
PASS Pasting text into editable <b>
PASS Pasting text in <i> into editable <b>
PASS Pasting 2 paragraphs
PASS Pasting 2 paragraphs into <b>
PASS Pasting 2 paragraphs whose text is bold
FAIL Pasting text in <b>: beforeinput assert_true: dataTransfer should have the copied HTML source expected true got false
PASS Pasting text in <b>: pasted result
FAIL Pasting text in <span>: beforeinput assert_true: dataTransfer should have the copied HTML source expected true got false
PASS Pasting text in <span>: pasted result
PASS Pasting text into editable <b>: beforeinput
FAIL Pasting text into editable <b>: pasted result assert_equals: text should be inserted into the editable <b> expected "<b>AabcB</b>" but got "<b>A</b>abc<b>B</b>"
FAIL Pasting text in <i> into editable <b>: beforeinput assert_true: dataTransfer should have the copied HTML source expected true got false
FAIL Pasting text in <i> into editable <b>: pasted result assert_equals: text should be inserted into the editable <b> without copied <i> expected "<b>AabcB</b>" but got "<b>A</b>abc<b>B</b>"
FAIL Pasting 2 paragraphs: beforeinput assert_true: dataTransfer should have the copied HTML source expected true got false
FAIL Pasting 2 paragraphs: pasted result assert_in_array: Each paragraph should be pasted as a line value "AabcB" not in array ["Aabc<br>defB", "A<br>abc<br>def<br>B"]
FAIL Pasting 2 paragraphs into <b>: beforeinput assert_true: dataTransfer should have the copied HTML source expected true got false
FAIL Pasting 2 paragraphs into <b>: pasted result assert_in_array: Each paragraph should be pasted as a line value "<b>A</b>abc<b>B</b>" not in array ["<b>Aabc<br>defB</b>", "<b>A<br>abc<br>def<br>B</b>"]
FAIL Pasting 2 paragraphs whose text is bold: beforeinput assert_true: dataTransfer should have the copied HTML source expected true got false
FAIL Pasting 2 paragraphs whose text is bold: pasted result assert_in_array: Each paragraph should be pasted as a line value "AabcB" not in array ["Aabc<br>defB", "A<br>abc<br>def<br>B"]
